Preparing for overhaul of GraphicsClass
